role dan permission complete ('maybe')

+++ app/Http/Controllers/PeminjamanMasterController.php
@@ -7,135 +7,62 @@ namespace App\Http\Controllers;
 
 
 use Illuminate\Http\Request;
-
 use Yajra\DataTables\DataTables;
 
-
-
 use Auth;
-
 use View;
-
 use Session;
-
 use App\User;
-
 use App\PeminjamanMaster;
-
-
-
 use App\PeminjamanPengecekan;
-
-
-
 class PeminjamanMasterController extends Controller
 
 {
-
-    
-
     public function index(){
-
-
-
+        $this->authorize('bukutanah.read');
         return view('peminjaman.peminjamanmaster.index');
-
     }
-
-
 
     public function cekpinjam(Request $request)
-
     {
-
-
-
+        $this->authorize('bukutanah.crud');
         $checklike = PeminjamanPengecekan::where('id', '=', $request->id)
-
             ->where('status_pinjam', '=', $request->status_pinjam)
-
             ->first();
 
-
-
         if ($checklike->status_pinjam == 0) {
-
             $data = PeminjamanPengecekan::find($request->id);
-
             $data->status_pinjam = 1;
-
             $data->update();
-
-
-
             Session::flash('info', 'BT tidak tersedia');
-
             return View::make('layouts/alerts');
-
-
-
         } else {
-
             $data = PeminjamanPengecekan::find($request->id);
-
             $data->status_pinjam = 0;
-
             $data->update();
-
-
-
             Session::flash('info', 'BT kembali tersedia');
-
             return View::make('layouts/alerts');
-
         }
-
     }
-
-
-
     public function cekpinjam2(Request $request)
-
     {
-
+        $this->authorize('bukutanah.crud');
         $checklike = PeminjamanPengecekan::where('id', '=', $request->id)
-
         ->where('status_pinjam2', '=', $request->status_pinjam2)
-
         ->first();
-
-
-
         if ($checklike->status_pinjam2 == 0) {
-
             $data = PeminjamanPengecekan::find($request->id);
-
             $data->status_pinjam2 = 1;
-
             $data->update();
-
-
-
             Session::flash('info', 'SU tidak tersedia');
-
             return View::make('layouts/alerts');
-
-
 
         } else {
-
             $data = PeminjamanPengecekan::find($request->id);
-
             $data->status_pinjam2 = 0;
-
             $data->update();
-
-
-
             Session::flash('info', 'SU kembali tersedia');
-
             return View::make('layouts/alerts');
-
         }
 
     }
@@ -145,6 +72,7 @@ class PeminjamanMasterController extends Controller
     public function apiPeminjamanMaster()
     {
         // $data = PeminjamanPengecekan::all();
+        $this->authorize('bukutanah.read');
         $data = PeminjamanPengecekan::orderBy('updated_at');
         return Datatables::of($data)
             ->addColumn('action', function ($data) {
@@ -167,21 +95,16 @@ class PeminjamanMasterController extends Controller
     }
 
     public function autoCompletePeminjaman(Request $request)
-
     {
-
+        $this->authorize('bukutanah.crud');
         $query = $request->get('term','');
-
         $dataptsl = PeminjamanMaster::where('idbukutanah', 'LIKE', '%' . $query . '%')->limit(20)->get();
-
         $data = [];
-
         foreach ($dataptsl as $d) {
 
             $data[] = array('value' => $d->idbukutanah . ' || Jenis Hak: ' . $d->jenis_hak . ' || No Hak: ' . $d->no_hak . ' || Desa: ' . $d->desa . ' || Kec: ' . $d->kecamatan, 'id' => $d->idbukutanah);
 
         }
-
         if (count($data)) {
 
             return $data;
@@ -191,18 +114,14 @@ class PeminjamanMasterController extends Controller
             return ['value' => 'Data tidak ada', 'id' => ''];
 
         }
-
     }
 
 
 
     public function autoCompletePeminjamanshow(Request $request)
-
     {
-
+        $this->authorize('bukutanah.crud');
         $id = $request->idbukutanah;
-
-
 
         $datas = PeminjamanMaster::where('idbukutanah', $id)->first();
 
@@ -219,59 +138,34 @@ class PeminjamanMasterController extends Controller
             'no_hak' => $datas->no_hak,
 
             'no_box' => $datas->no_box,
-
         );
-
         $row_set[] = $data;
-
         return $return = json_encode($row_set);
 
     }
 
-
-
-
-
     public function edit($id)
-
     {
-
+        $this->authorize('bukutanah.read');
         return PeminjamanPengecekan::find($id);
-
     }
 
-
-
     public function update(Request $request, $id)
-
     {
-
+        $this->authorize('bukutanah.crud');
         $data = PeminjamanPengecekan::find($id);
-
         $data->no_box = $request['no_box'];
-
         $data->idbukutanah = $request['idbukutanah'];
-
         $data->jenis_hak = $request['jenis_hak'];
-
         $data->no_hak = $request['no_hak'];
-
         $data->desa = $request['desa'];
-
         $data->kecamatan = $request['kecamatan'];
-
         $data->ruang = $request['ruang'];
-
         $data->rak = $request['rak'];
-
         $data->baris = $request['baris'];
-
         $data->status_pinjam = $request['status_pinjam'];
-
         $data->status_pinjam2 = $request['status_pinjam2'];
-
         $data->created_by = $request['created_by'];
-
         $data->update();
 
         Session::flash('info', 'Data Berhasil Diubah');
@@ -280,36 +174,21 @@ class PeminjamanMasterController extends Controller
 
     }
 
-
-
     public function destroy($id)
-
     {
-
+        $this->authorize('bukutanah.crud');
         PeminjamanPengecekan::destroy($id);
-
         Session::flash('info', 'Data Berhasil Dihapus');
 
         return View::make('layouts/alerts');
-
     }
 
-
-
     public function store(Request $request)
-
     {
-
+        $this->authorize('bukutanah.crud');
         PeminjamanPengecekan::create($request->except(['_method', '_token']));
-
-
-
         Session::flash('info', 'Data Baru Berhasil Ditambah');
-
         return View::make('layouts/alerts');
-
-
-
     }
